Use a new property VersionLabel for mutating suffix
If we use the built-in VersionSuffix, any referenced project that doesn't import our directory targets (i.e. a submodule) will fail to build because the version suffix will be invalid for them.

So use an entirely new property instead, which will be ignored by other projects, that we can modify and set to the VersionSuffix built-in property instead.