Build and update changelog on new releases
This independent workflow runs whenever a release is released (created as a non-prerelease up-front or changed afterwards) and updates the public changelog in the repo, which is linked from released nuget packages.

We don't update the changelog on release creation/publishing because we'll use that as a validation step instead (create as prerelease, have notes+artifacts populate).

Fixes #4