Update a release notes and artifacts on publishing
Instead of the current flow where we need to explicitly tag via git cli to trigger our workflows, make it so that we create a release on GitHub and that triggers the changelog generation and updates the release notes automatically.

We should also build the project and upload as artifacts, so that they can be validated, before releasing (as non-prerelease).

Since there is no way to force the release created event without actually publishing it, we allow the project owner to create the release as prerelease, causing all the above to trigger, and only the final release (when changing the release from prerelease to non-prerelease) is the one that builds and pushes to nuget.org.

Fixes #5