Automatically use higher-permissions PAT for push
The GH CLI already automatically switches to a GH_TOKEN instead of GITHUB_TOKEN envvar if available, since many operations enabled by the CLI require a token with more scopes than the default one available for actions.

We apply that same concept to our changelog push, by automatically using GH_TOKEN if present, but defaulting to GITHUB_TOKEN if not. This allows repos with main branch protections to still successfully commit.

Need to be aware of potential endless loop of workflows in this case though, but we only commit from releases, so no risk there.