🔄 Automate dotnet-file sync
Every night, run `dotnet file sync` and create a PR as needed,
against the same branch so that if it's not merged, it's kept
as an ongoing PR until merged.

The body of the PR is the changelog generated by the sync operation,
so it's easy to explore incoming changes.