Automatically set/include icon.png and readme.md
To reduce the ceremony in producing great looking packages,
we expand a bit the heuristics for picking up the package icon
and add the readme too, as follows:

* If an icon.png exists alongside the project file, assume it will
  be the package icon: add/update its None item and pack appropriately
* If a project-level icon.png doesn't exist, apply the same behavior
  but at the src (alognside the Directory.Build.* file) level.

Do the same for a file named `readme.md` at both levels.

The reason to use a src-level readme as opposed to a repository
root one, is that the repository readme might include information
that's intended for repository/sources consumption (such as how to
clone, contribute, report bugs, etc.), whereas the package readme
might be different. For that reason, the project-level icon/readme
also prevail a generic src-level one.