Fix test filter for theory tests
Theories contain additional characters in the test display name,
including whitespaces and a rendering of the data elements for each
test, such as `Devlooped.DocumentRepositoryTests.DocumentEndToEnd(serializer: DocumentSerializer { })`.

The existing regex would truncate the string at the whitespace,
resulting in `Devlooped.DocumentRepositoryTests.DocumentEndToEnd(serializer:`
which then wouldn't match any tests, making it appear that the test
run was otherwise successful.

This fixes that by properly excluding everything after the test name
and deduplicating the resulting list (since the presence of theories
would result in duplicates now). This means that we run all theories
for a single instance failure, but the alternative of filtering for
the full test by display name (including theory data) in the filter
didn't work.