Fix incremental build issue with package files and doc file
The doc file for a project is not re-generated unless there are new APIs. This means it can be older than a package file (i.e. a .targets file) in the project, and remain so for a long time across many builds.

This slows down builds unnecessarily, so remove. If package really needs a new build, you can always rebuild before Pack.