Use a better CI version number
By making the last number the build number, we ensure we have ever-incrementing numbers, even across branches. This makes it easier to have wildcard versions too, and the prerelease label becomes just the branch name (or tag name).

We also simplify the actual build/pack calls by using envvars to set both the VersionPreffix and VersionLabel. We don't use VersionSuffix for the latter since that would be confusing with the built-in property (actually populated from the VersionLabel) which does no additional processing. We do, for our labels (by setting PR numer, removing 'v' in a tag, and so on).