Extend grace period to unknown status too
Use the existing Funding.Grace value also as the value for an install-time (optional) grace period where no warnings will be reported.

In order for the install-time check to be performed, the package must provide an item `SponsorablePackageId` with the package id containing the analyzer.

The install check can be skipped (for testing purposes) by adding a compiler visible property, like so:

```xml
  <!-- Force immediate reporting of status, no install-time grace period -->
  <PropertyGroup>
    <SponsorLinkNoInstallGrace>true</SponsorLinkNoInstallGrace>
  </PropertyGroup>
  <ItemGroup>
    <CompilerVisibleProperty Include="SponsorLinkNoInstallGrace" />
  </ItemGroup>
```

This allows testing the behavior of the unknown status warning right after installing the package.

Note that if no install time can be determined, the unknown status will always be reported.