Fix roles checking from new identity-based token handler
The roles in JWT aren't mapped to the namespaced role, and we want to keep it that way for consistency with the bare JWT `roles`.

By extending the ClaimsPrincipal we return from manifest reading, we can maintain compatibility with minimal changes.

See https://github.com/Azure/azure-functions-host/issues/3898, for example.