Make sure we report only once per product for entire solution
Since many projects can use the same package in a solution, it makes no sense to report multiple times the same diagnostic asking for funding. It would just be super annoying.

So make it process-wide and singleton. We use a mutex+memory-mapped file which lives alongside the whole process, since individual projects may be built separately and cause duplicate reporting even if we use use static or AppDomain-stored state.