Attempt to get necessary permissions for default token
If the repo or organization has the right setting, this should make this Just Work without having to set up a custom GH_TOKEN.

The relevant setting is under `Settings > Actions > General`. Under "Workflow permissions", check "Allow GitHub Actions to create and approve pull requests". This setting needs to be enabled for the repository or organization in order for the default token to work.