Move dotnet setup to composite action
Also introduces a repo custom property we can set to request a different set of dotnet versions easily without having to modify the workflow itself, making it more future-proof.

For now, since --allow-downgrade used by the sleet step requires .NET 9.x to be explicitly setup to get latest, this will run always on all our repos since we set the DOTNET prop to `9.x` by default. When dotnet 10 ships, this will likely not be needed anymore and we'll remove it.